// src/common/enums/lesson-messages.enum.ts
export enum LessonMessages {
    NOT_FOUND = 'Lesson not found',
    ALREADY_EXISTS = 'Lesson already exists',
    CREATED = 'Lesson created successfully',
    UPDATED = 'Lesson updated successfully',
    DELETED = 'Lesson deleted successfully',

    UNAUTHORIZED_ACCESS = 'Unauthorized access',
    INVALID_CREDENTIALS = 'Invalid credentials',
}